refactor(pt_expt): refactor frozen model with dpmodel base and torch_module
- Create dpmodel FrozenModel (NativeOP + BaseModel) with all delegation
  methods, so pt_expt can inherit instead of duplicating
- Rewrite pt_expt FrozenModel to use @torch_module wrapping dpmodel class
- Override __init__ to handle .pte files natively via serialize_from_file,
  fall back to generic backend detection for other formats
- Override serialize() to delegate directly to inner model (unlike pt
  which must reconstruct from model_def_script due to opaque ScriptModule)
- Add pt_expt support to frozen consistency test using BaseModel as
  pt_expt_class (same pattern as pt)
- Guard setUpModule model generation with backend availability checks
